Fix CollectSummaries.
