+++ tests/link-dce-failing/stdlib-link-map-2/Test.scala
