+++ tests/link-dce-failing/stdlib-link-patmatch-4/Test.scala
