+++ tests/link-dce-failing/stdlib-link-vector-4/Test.scala
