+++ tests/link-dce-stdlib/stdlib-link-array-mkString/Test.scala
@@ -1,7 +1,7 @@
 import scala.annotation.internal
 
 object Test {
-  @internal.link.CallGraphBounds(reachableClasses = 132, classesWithReachableMethods = 15, reachableMethods = 74)
+  @internal.link.CallGraphBounds(reachableClasses = 332, classesWithReachableMethods = 284, reachableMethods = 573)
   def main(args: Array[String]): Unit = {
     System.out.println(Array(1, 2, 3).mkString(", "))
   }
