+++ tests/link-dce-stdlib/stdlib-link-break/Test.scala
@@ -1,7 +1,7 @@
 import scala.annotation.internal
 
 object Test {
-  @internal.link.CallGraphBounds(reachableClasses = 25, classesWithReachableMethods = 7, reachableMethods = 9)
+  @internal.link.CallGraphBounds(reachableClasses = 70, classesWithReachableMethods = 17, reachableMethods = 29)
   def main(args: Array[String]): Unit = {
     import scala.util.control.Breaks._
     breakable {
