+++ tests/link-dce-stdlib/stdlib-link-patmatch-3/Test.scala
