+++ tests/link-dce-stdlib/stdlib-link-predef-println-1/Test.scala
@@ -1,0 +1,8 @@
+import scala.annotation.internal
+
+object Test {
+  @internal.link.CallGraphBounds(reachableClasses = 104, classesWithReachableMethods = 79, reachableMethods = 145)
+  def main(args: Array[String]): Unit = {
+    println(42)
+  }
+}
