--- tests/link-dce-stdlib/stdlib-link-predef-println-1/stdlib-link-predef-println-1.scala
@@ -1,8 +1,0 @@
-import scala.annotation.internal
-
-object Test {
-  @internal.link.CallGraphBounds(reachableClasses = 91, classesWithReachableMethods = 8, reachableMethods = 57)
-  def main(args: Array[String]): Unit = {
-    println(42)
-  }
-}
