+++ tests/link-dce-stdlib/stdlib-link-predef-println-2/Foo.scala
@@ -1,0 +1,3 @@
+class Foo {
+  override def toString: String = "foo"
+}
