+++ tests/link-dce-stdlib/stdlib-link-set-0/Test.scala
@@ -1,7 +1,7 @@
 import scala.annotation.internal
 
 object Test {
-  @internal.link.CallGraphBounds(reachableClasses = 49, classesWithReachableMethods = 7, reachableMethods = 7)
+  @internal.link.CallGraphBounds(reachableClasses = 53, classesWithReachableMethods = 42, reachableMethods = 51)
   def main(args: Array[String]): Unit = {
     scala.collection.immutable.Set.empty[Int]
   }
