+++ tests/link-dce-stdlib/stdlib-link-trycatch-unapply/Bar.scala
@@ -1,0 +1,1 @@
+class Bar extends Exception("BarException")
