+++ tests/link-dce-stdlib/stdlib-link-trycatch-unapply/Foo.scala
@@ -1,0 +1,1 @@
+class Foo(e: Throwable) extends Exception(e)
