+++ tests/link-dce-stdlib/stdlib-link-trycatch-unapply/Test.scala
@@ -1,14 +1,12 @@
 import scala.annotation.internal
 
 object Test {
-  class Bar extends Exception("BarException")
-  class Foo(e: Throwable) extends Exception(e)
-  
+
   object CausedBy {
     def unapply(e: Throwable): Option[Throwable] = Option(e.getCause)
   }
 
-  @internal.link.CallGraphBounds(reachableClasses = 33, classesWithReachableMethods = 14, reachableMethods = 20)
+  @internal.link.CallGraphBounds(reachableClasses = 40, classesWithReachableMethods = 24, reachableMethods = 42)
   def main(args: Array[String]): Unit = {
     try {
         throw new Foo(new Bar)
