+++ tests/link-dce-stdlib/stdlib-link-varargs-AnyRef/Test.scala
@@ -1,9 +1,9 @@
 import scala.annotation.internal
 
 object Test {
-  @internal.link.CallGraphBounds(reachableClasses = 123, classesWithReachableMethods = 13, reachableMethods = 62)
+  @internal.link.CallGraphBounds(reachableClasses = 304, classesWithReachableMethods = 263, reachableMethods = 496)
   def main(args: Array[String]): Unit = {
-    foo()
+    foo(new Object)
   }
 
   def foo(ns: AnyRef*): Unit = {
