+++ tests/link-dce-stdlib/stdlib-link-varargs-int/Test.scala
@@ -1,7 +1,7 @@
 import scala.annotation.internal
 
 object Test {
-  @internal.link.CallGraphBounds(reachableClasses = 123, classesWithReachableMethods = 13, reachableMethods = 62)
+  @internal.link.CallGraphBounds(reachableClasses = 303, classesWithReachableMethods = 263, reachableMethods = 494)
   def main(args: Array[String]): Unit = {
     foo(1, 2, 3)
   }
