+++ tests/link-dce-stdlib/stdlib-link-varargs-poly1/Test.scala
@@ -1,14 +1,14 @@
 import scala.annotation.internal
 
 object Test {
-  @internal.link.CallGraphBounds(reachableClasses = 123, classesWithReachableMethods = 13, reachableMethods = 62)
+  @internal.link.CallGraphBounds(reachableClasses = 311, classesWithReachableMethods = 271, reachableMethods = 511)
   def main(args: Array[String]): Unit = {
     foo(1)
   }
 
   def bar[T](x: T) = foo(x)
 
-  def foo[U](ns: U*): Unit = {
+  def foo(ns: Any*): Unit = {
     System.out.println(42)
   }
