+++ tests/link-dce-stdlib/stdlib-link-vector-1/Test.scala
@@ -1,7 +1,7 @@
 import scala.annotation.internal
 
 object Test {
-  @internal.link.CallGraphBounds(reachableClasses = 57, classesWithReachableMethods = 7, reachableMethods = 9)
+  @internal.link.CallGraphBounds(reachableClasses = 62, classesWithReachableMethods = 51, reachableMethods = 67)
   def main(args: Array[String]): Unit = {
     scala.collection.immutable.Vector.empty
   }
